BUGFIX: when the AbstractDSmeasure was constructed w/o Ops (which
is the case of most DS measures), but then its calculate() was called
with some _cache and the re-created ImgQualityDataCache obj was built
with the Ops from the DSmeasure (with no Ops) and was not re-using
the Ops from the given _cache (which had Ops). Changed to give "preference"
to using Ops from the given _cache, and only if that is not possible
it attempts to use Ops from the AbstractDSmeasure...

Changed ImgQualityDataCache() c'tor to avoid confusion which Ops will be taken.